backend/engine/templates/generated/b597a288-2647-4952-a5ed-78d010376e29/requirements.txt
fastmcp>=0.2.0
pykiteconnect==3.9.3
pydantic>=1.9.0,<2.0.0
python-dotenv>=0.19.0
uvicorn[standard]>=0.15.0
httpx>=0.20.0
